[M4] Feedback - Introdice SetUpFixture and incapsulate LoginWithCredentials for default test user
